+++ tests/src/test/java/de/sovity/edc/e2e/Ms8ConnectorMigrationTest.java
@@ -45,8 +45,8 @@ import static org.assertj.core.api.Assertions.assertThat;
  */
 class Ms8ConnectorMigrationTest {
 
-    private static final String PROVIDER_PARTICIPANT_ID = "provider";
-    private static final String CONSUMER_PARTICIPANT_ID = "consumer";
+    private static final String PROVIDER_PARTICIPANT_ID = "example-provider";
+    private static final String CONSUMER_PARTICIPANT_ID = "example-connector";
 
     @RegisterExtension
     static EdcExtension providerEdcContext = new EdcExtension();
@@ -150,6 +150,7 @@ class Ms8ConnectorMigrationTest {
         assertThat(providerTransfer.getAssetName()).isEqualTo("First Asset");
         assertThat(providerTransfer.getContractAgreementId()).isEqualTo("Zmlyc3QtY2Q=:Zmlyc3QtYXNzZXQtMS4w:MjgzNTZkMTMtN2ZhYy00NTQwLTgwZjItMjI5NzJjOTc1ZWNi");
         assertThat(providerTransfer.getCounterPartyConnectorEndpoint()).isEqualTo(endpoint(consumerConnector));
+        assertThat(providerTransfer.getCounterPartyParticipantId()).isEqualTo(consumerConnector.getParticipantId());
         assertIsEqualOffsetDateTime(providerTransfer.getCreatedDate(), EdcDateUtils.utcMillisToOffsetDateTime(1695208010855L));
         assertThat(providerTransfer.getDirection()).isEqualTo(ContractAgreementDirection.PROVIDING);
         assertThat(providerTransfer.getErrorMessage()).isNull();
@@ -176,6 +177,7 @@ class Ms8ConnectorMigrationTest {
         assertThat(consumerTransfer.getAssetName()).isEqualTo("first-asset-1.0");
         assertThat(consumerTransfer.getContractAgreementId()).isEqualTo("Zmlyc3QtY2Q=:Zmlyc3QtYXNzZXQtMS4w:MjgzNTZkMTMtN2ZhYy00NTQwLTgwZjItMjI5NzJjOTc1ZWNi");
         assertThat(consumerTransfer.getCounterPartyConnectorEndpoint()).isEqualTo(endpoint(providerConnector));
+        assertThat(consumerTransfer.getCounterPartyParticipantId()).isEqualTo(providerConnector.getParticipantId());
         assertIsEqualOffsetDateTime(consumerTransfer.getCreatedDate(), EdcDateUtils.utcMillisToOffsetDateTime(1695208008652L));
         assertThat(consumerTransfer.getDirection()).isEqualTo(ContractAgreementDirection.CONSUMING);
         assertThat(consumerTransfer.getErrorMessage()).isNull();